Refactored ChessBoard
Instead of copying the FEN string over and using that in the draw
function I now use the string to construct a map of pieces.
This map is keyed on STSquare.

This enables me to use this map to find out which piece needs to
be dragged.

Dragging virtually complete. Need to despatch to the engine once the
drag is completed.