Last of the GUI changes for the moment.
Reduced the amount of functions called from the logic to a minimum.
The idea is to have an interface into the engine with the engine not
calling back at all. Thus the app pointer in the engine should not
be needed.
Not quite there yet, three more functions to go.