Consolidated passing around colour data to the STColour struct.
Instead of using getters to extract individual colour components
and passing each component down the line, pass a complete struct
down and get values direct.

We can do this because we are ensuring that the struct is immutable.
Always make sure you copy the struct through passing by value.

The calls init_colours and is_colour use references as parameters.
is_colour is declared const and thus keeps the struct immutable.
init_colour is used as a substitute for the constructor.
Further refactoring is required. This function needs to disappear.
The struct should only be set by the constructor.