Refactored virtual function names to clearly indicate their function
Also fixed a subtle bug in set_colours (now setup). The previous
code looked "back" and tried to update the ColourChooser parent class.
It is supposed to look forward and manipulate external objects
(DialogColours) and set it up for custom operations.

The same in the result function, which as a result of this backward
gazing was doing a self assignment on the STColour variable.