BUGFIX: level change was executed twice
When you select a different level GTK creates a de-selection message
on the old level and a selection message on the new level.
As I was not testing on this the action got executed twice.
Added a current_level variable to the controller to enable us
to disregard messages where the level does not change.
Is there a possibility to interrogate GTK to find out if this is a
de-selection or selection message?