Changed drawing method
Instead of repainting every in on_draw, I now build up image overlays
which can be bit blitted in the on_draw function.
Apart from the fact that the lower layers now have a simpler data
structure to deal with (just an image) there is a slight performance
enhancement. We only interpret incoming data once and generate our
overlay for painting, instead of interpreting the data each time
the screen is redrawn.