Lexer: Simplify and flesh out the lexer
Now that the tokeniser is handling line comments we can greatly
simplify the lexer. I also removed the dependency on the ChessGame
object so the lexer can be more easily tested in isolation.

This affects the function signature of the parse_tag_pair_section
function.

I also added the parse_movetext_section function which eventually
will parse all the moves.