Console version
I found Unicode characters for chess pieces on the net and wondered if I
could make a chess representation on the console just outputting text
characters.
It got a bit out of hand and after the representation was finished I
started adding logic to it.
Currently the engine generates moves and selects a random move on the
AI's turn.