BUGFIX
If dragging starts outside the board (ie from the edit bitmap) the start
square is -1. This is not a valid index into the array.
The same goes for the end square. If the mouse is released outside the
board, the index is not valid.
NB. commented out the demo code as the new Ply class needs constructor
parameters.