PrintSAN
What sense does it make for the caller of print_SAN to know that it
needs a list of valid ply's. Pass in the game state and print_SAN can
generate it if need be. (Saves some execution cycles as well if print_SAN
can figure it out without the list.)